refact: confirm account box component
