+++ src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
 import { useUser } from '../contexts/UserContext';
@@ -6,56 +6,12 @@ import { useAuth } from '../contexts/AuthContext';
 import Navbar from '../components/Navbar';
 
 import '../style/Home.css';
+import ConfirmAccountBox from '../components/ConfirmAccountBox';
 
 function Home() {
   const navigate = useNavigate();
   const { accessToken } = useAuth();
   const { userData, setUserData } = useUser();
-
-  const [email, setEmail] = useState('');
-  const [successMessage, setSuccessMessage] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleEmailSubmit = async (e) => {
-    e.preventDefault();
-
-    setIsLoading(true);
-    setSuccessMessage('');
-    setErrorMessage('');
-
-    try {
-      await axios.patch(
-        'http://localhost:3000/auth/account/resend-token',
-        { email },
-        {
-          headers: {
-            Authorization: `Bearer ${accessToken}`,
-          },
-        },
-      );
-
-      setUserData((prevUserData) => ({
-        ...prevUserData,
-        email,
-      }));
-
-      setSuccessMessage(
-        `Email para confirmação de conta enviado para ${email}`,
-      );
-    } catch (error) {
-      setErrorMessage('Erro ao enviar novo email para confirmação de conta');
-
-      const axiosErrorMessage = error.response.data.error.message;
-
-      if (axiosErrorMessage === 'The new email provided is already in use') {
-        setErrorMessage(`O email informado (${email}) já está em uso.`);
-      }
-    }
-
-    setEmail('');
-    setIsLoading(false);
-  };
 
   useEffect(() => {
     const fetchUserData = async () => {
@@ -81,49 +37,11 @@ function Home() {
 
       <div className="content-container">
         {userData.status !== 'ACTIVE' && (
-          <div id="confim-account-box">
-            <p id="message-box">
-              Sua conta ainda não foi confirmada. Enviamos um email para
-              {' '}
-              <strong>{userData.email}</strong>
-              , basta clicar no link de
-              confirmação para obter acesso total na plataforma!
-            </p>
-
-            <form onSubmit={handleEmailSubmit}>
-              <div className="container-form">
-                <label htmlFor="email">
-                  <input
-                    name="email"
-                    type="email"
-                    id="email"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    required
-                    placeholder={userData.email}
-                  />
-                </label>
-              </div>
-
-              {successMessage && (
-                <div className="success-msg">
-                  <p>{successMessage}</p>
-                </div>
-              )}
-
-              {errorMessage && (
-                <div className="error-msg">
-                  <p>{errorMessage}</p>
-                </div>
-              )}
-
-              <div className="inputs-buttons">
-                <button type="submit" disabled={isLoading}>
-                  {isLoading ? 'Enviando email...' : 'Enviar novo email'}
-                </button>
-              </div>
-            </form>
-          </div>
+        <ConfirmAccountBox
+          accessToken={accessToken}
+          userData={userData}
+          setUserData={setUserData}
+        />
         )}
       </div>
     </section>
